+++ .gitignore
@@ -4,5 +4,6 @@ coverage
 dist
 xcuserdata
 .next
+.data.json
 *.tsbuildinfo
 *.log
