+++ packages/site/src/components/controls/SegmentedControl.tsx
@@ -7,8 +7,8 @@ import {
 } from "@radix-ui/react-toggle-group"
 import { HTMLMotionProps, LayoutGroup, motion } from "framer-motion"
 import { ReactChild, memo } from "react"
-import { springy } from "../transitions"
-import { mergeProps } from "../utils/merge-props"
+import { springy } from "../../transitions"
+import { mergeProps } from "../../utils/merge-props"
 
 interface Props extends Omit<ToggleGroupSingleProps, "type"> {
   /**
